Refactor common code from GithubCalls to abstract class, move project extensions to classes where they are used.
