+++ kmmbridge/src/main/kotlin/versionmanager/TimestampVersionManager.kt
@@ -13,7 +13,6 @@
 
 package co.touchlab.faktory.versionmanager
 
-import co.touchlab.faktory.internal.alwaysWriteGitTags
 import co.touchlab.faktory.internal.writeGitTagVersion
 import org.gradle.api.Project
